Move interpreter traversal into mech-program
