+++ src/runtime/Cargo.toml
@@ -150,7 +150,6 @@ notify = ["dep:notify"]
 [dependencies]
 blake3 = "1.8.5"
 mech-program = { version = "0.3.5", default-features = false }
-mech-interpreter = { version = "0.3.5", default-features = false }
 mech-core = { version = "0.3.5", default-features = false }
 mech-syntax = { version = "0.3.5", default-features = false }
 serde = { version = "1.0.228", optional = true }
